requirements.txt
beautifulsoup4==4.11.1
Django==3.2.10
line_bot_sdk==2.3.0
lxml==4.6.3
PyMySQL==1.0.2
requests==2.26.0
